Fix error handling when token is chunked
In some cases, token returned exceeds 2048 bytes and is returned as
chunked, in which case Content-length is set to 0.

`getInputStream()` handles chunked contents gracefully, but the check
for the Content-length to 0 breaks the use-case.

Also, if the HTTP ersponse code isn't 200, we want to log all of the
output instead of chopping it off at 1024 bytes.

Fixes #63